refactor(processor): extract pdf byte helpers
Move content fingerprinting and truncated-file hints out of the main processor so document parsing stays focused on the pipeline flow.
